refactor: move `web.date.util` inside `@web/common/utils/dates` for organization
