+++ packages/web/src/views/Calendar/components/Draft/sidebar/hooks/useSidebarActions.ts
@@ -20,13 +20,13 @@ import {
 } from "@web/common/constants/web.constants";
 import { Coordinates } from "@web/common/types/util.types";
 import { isEventFormOpen } from "@web/common/utils";
-import { assembleDefaultEvent } from "@web/common/utils/event.util";
-import { getX } from "@web/common/utils/grid.util";
 import {
   computeCurrentEventDateRange,
   computeRelativeEventDateRange,
   getDatesByCategory,
-} from "@web/common/utils/web.date.util";
+} from "@web/common/utils/dates/web.date.util";
+import { assembleDefaultEvent } from "@web/common/utils/event.util";
+import { getX } from "@web/common/utils/grid.util";
 import {
   selectDraft,
   selectDraftActivity,
