+++ packages/web/src/views/Calendar/components/Event/Grid/GridEvent/GridEvent.tsx
@@ -12,11 +12,11 @@ import {
   ZIndex,
 } from "@web/common/constants/web.constants";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
+import { getTimesLabel } from "@web/common/utils/dates/web.date.util";
 import { isOptimisticEvent } from "@web/common/utils/event.util";
 import { getLineClamp } from "@web/common/utils/grid.util";
 import { isRightClick } from "@web/common/utils/mouse/mouse.util";
 import { getEventPosition } from "@web/common/utils/position.util";
-import { getTimesLabel } from "@web/common/utils/web.date.util";
 import { Flex } from "@web/components/Flex";
 import {
   AlignItems,
