+++ packages/web/src/views/Calendar/components/Grid/MainGrid/MainGrid.tsx
@@ -2,10 +2,10 @@ import { Dayjs } from "dayjs";
 import React, { FC, MouseEvent, MutableRefObject } from "react";
 import { Categories_Event } from "@core/types/event.types";
 import { ID_GRID_MAIN } from "@web/common/constants/web.constants";
+import { getHourLabels } from "@web/common/utils/dates/web.date.util";
 import { assembleDefaultEvent } from "@web/common/utils/event.util";
 import { getX } from "@web/common/utils/grid.util";
 import { isRightClick } from "@web/common/utils/mouse/mouse.util";
-import { getHourLabels } from "@web/common/utils/web.date.util";
 import { selectIsDrafting } from "@web/ducks/events/selectors/draft.selectors";
 import { draftSlice } from "@web/ducks/events/slices/draft.slice";
 import { useAppDispatch, useAppSelector } from "@web/store/store.hooks";
