+++ packages/web/src/views/Forms/EventForm/DateControlsSection/DateTimeSection/TimePicker/TimePickers.tsx
@@ -2,12 +2,12 @@ import React, { FC, useState } from "react";
 import { Schema_Event } from "@core/types/event.types";
 import { SelectOption } from "@web/common/types/component.types";
 import { Option_Time } from "@web/common/types/util.types";
-import { shouldAdjustComplimentTime } from "@web/common/utils/datetime/web.datetime.util";
 import {
   getTimeOptionByValue,
   getTimeOptions,
   mapToBackend,
-} from "@web/common/utils/web.date.util";
+} from "@web/common/utils/dates/web.date.util";
+import { shouldAdjustComplimentTime } from "@web/common/utils/datetime/web.datetime.util";
 import { AlignItems } from "@web/components/Flex/styled";
 import { StyledTimeFlex } from "../styled";
 import { TimePicker } from "./TimePicker";
