+++ packages/web/src/views/Forms/SomedayEventForm/SomedayEventForm.test.tsx
@@ -12,7 +12,7 @@ import {
 } from "@core/types/event.types";
 import { createMockStandaloneEvent } from "@core/util/test/ccal.event.factory";
 import { render } from "@web/__tests__/__mocks__/mock.render";
-import { setupDraftState } from "@web/__tests__/utils/grid.util/draft.util";
+import { setupDraftState } from "@web/__tests__/utils/state/draft.state.util";
 import { Schema_WebEvent } from "@web/common/types/web.event.types";
 import { EventApi } from "@web/ducks/events/event.api";
 import { deleteEventSlice } from "@web/ducks/events/slices/event.slice";
